refactor: qol.hpp => utils.hpp
